refactor(fast_io): decompose io_uring/registered_buffers into submodules
The monolithic 1607-line file mixed the registration coordinator,
slot allocator, telemetry, batch submit helpers, and tests. Split
into focused submodules under registered_buffers/ to bring each
file under the LoC audit cap and clarify ownership of each concern.

No public API changes - external imports (file_reader, file_writer,
shared_ring, integration tests) continue to resolve through the
re-export surface in mod.rs.
