--- crates/fast_io/src/io_uring/registered_buffers.rs
@@ -1,1607 +1,0 @@
-//! Page-aligned buffer registration for io_uring `READ_FIXED`/`WRITE_FIXED` operations.
-//!
-//! Registered buffers avoid kernel-userspace address translation on every SQE by
-//! pinning a set of fixed buffers via `IORING_REGISTER_BUFFERS`. The kernel maps
-//! these buffers once at registration time, eliminating per-op `get_user_pages()`
-//! calls - a significant win for high-throughput I/O.
-//!
-//! # Buffer lifecycle
-//!
-//! 1. **Allocate** - page-aligned buffers via [`std::alloc::alloc`] with proper layout.
-//! 2. **Register** - pass `iovec` array to `submitter.register_buffers()`.
-//! 3. **Checkout** - callers acquire a slot index for use with `ReadFixed`/`WriteFixed`.
-//! 4. **Return** - callers release the slot back to the free list.
-//! 5. **Drop** - frees user-side memory. Kernel-side unregistration happens
-//!    implicitly when the ring fd is closed; callers may also invoke
-//!    [`RegisteredBufferGroup::unregister`] explicitly while the ring is alive.
-//!
-//! # Drop ordering and the ring fd
-//!
-//! [`RegisteredBufferGroup`] does not hold a reference to the [`RawIoUring`]
-//! instance it was registered with. This is intentional: the kernel
-//! automatically releases the pinned user pages when the ring fd is closed
-//! (see `io_uring_register(2)` and `fs/io_uring.c:io_sqe_buffers_unregister`).
-//!
-//! Owners of both a `RawIoUring` and a `RegisteredBufferGroup` (such as
-//! `IoUringReader` and `IoUringWriter`) MUST declare the ring field BEFORE
-//! the `RegisteredBufferGroup` field. Rust drops fields in declaration
-//! order, so this ensures:
-//!
-//! 1. `RawIoUring::Drop` closes the ring fd first, releasing the kernel's
-//!    pinning of the registered buffer pages.
-//! 2. `RegisteredBufferGroup::Drop` then deallocates the user-side memory
-//!    backing those buffers.
-//!
-//! Reversing this order (group before ring) would still be sound because
-//! `Drop` only deallocates user memory and never touches the ring; the
-//! kernel would still hold the pinning until the ring fd later closes.
-//! However, the documented ordering matches the implementation in
-//! [`super::file_reader`] and [`super::file_writer`].
-//!
-//! # Why Drop does not call `unregister_buffers`
-//!
-//! Calling `submitter.unregister_buffers()` from `Drop` would require the
-//! group to hold a reference to the ring. That introduces lifetime coupling
-//! and makes it impossible for the ring to be dropped first - which is the
-//! natural ordering when the ring owns the group. Instead we rely on the
-//! kernel's automatic cleanup on ring fd close, and expose
-//! [`RegisteredBufferGroup::unregister`] for callers that want deterministic
-//! cleanup while keeping the ring alive (e.g., to register a new buffer set).
-//!
-//! # Panic safety
-//!
-//! `Drop` performs only `std::alloc::dealloc` calls, which do not panic when
-//! given a layout that matches the original allocation. This makes the impl
-//! safe during stack unwinding: a panic in user code that drops a
-//! `RegisteredBufferGroup` will not trigger a double-panic abort.
-//!
-//! # Process termination
-//!
-//! On `SIGKILL` or other forced exits, neither `Drop` nor any userspace
-//! cleanup runs, but the kernel reclaims both the ring fd and the registered
-//! buffer pages as part of normal process teardown. No leak occurs.
-//!
-//! # Kernel limits
-//!
-//! The maximum number of registered buffers is typically 1024 (kernel-dependent).
-//! Registration of more than the kernel supports returns `EINVAL` or `ENOMEM`.
-
-use std::alloc::{self, Layout};
-use std::io;
-use std::ptr;
-use std::sync::atomic::{AtomicU64, Ordering};
-
-use io_uring::IoUring as RawIoUring;
-
-/// Maximum number of buffers that can be registered with io_uring.
-///
-/// The kernel typically allows up to 1024 registered buffers. We cap at this
-/// limit to avoid kernel rejections.
-pub(super) const MAX_REGISTERED_BUFFERS: usize = 1024;
-
-/// A group of page-aligned buffers registered with an io_uring instance.
-///
-/// Each buffer is allocated with page alignment for optimal DMA and zero-copy
-/// performance. The group tracks which buffer indices are available via an
-/// atomic bitset, enabling lock-free checkout/return from multiple threads
-/// (though io_uring submission itself is single-threaded).
-///
-/// # Safety invariants
-///
-/// - All buffer pointers remain valid and pinned until `drop()`.
-/// - The `iovec` array passed to `register_buffers` references memory owned by
-///   this struct. The kernel holds references to these pages until
-///   `unregister_buffers` is called.
-/// - Buffer indices returned by `checkout()` are guaranteed to be in-bounds.
-///
-/// # Telemetry
-///
-/// The group records `total_acquires` and `total_misses` counters, both
-/// bumped on every [`checkout`](Self::checkout). A miss is a checkout that
-/// returns `None` because every slot is in use, which forces the caller
-/// to fall back to non-registered I/O. The counters are exposed via
-/// [`stats`](Self::stats) and feed the adaptive sizing design described
-/// in `docs/audits/io-uring-adaptive-buffer-sizing.md`.
-pub struct RegisteredBufferGroup {
-    /// Raw pointers to page-aligned buffer memory.
-    buffers: Vec<*mut u8>,
-    /// Layout used for each buffer allocation (for deallocation).
-    layout: Layout,
-    /// Size of each individual buffer in bytes.
-    buffer_size: usize,
-    /// Number of buffers in the group.
-    count: usize,
-    /// Atomic bitset tracking which buffer indices are free (1 = free, 0 = in use).
-    /// Supports up to 64 buffers per word. Multiple words for larger counts.
-    free_bitset: Vec<AtomicU64>,
-    /// Total number of `checkout` calls (whether they succeeded or not).
-    total_acquires: AtomicU64,
-    /// Number of `checkout` calls that returned `None` because every slot
-    /// was in use - a forced fallback to non-registered I/O.
-    total_misses: AtomicU64,
-}
-
-// SAFETY: The raw pointers point to memory exclusively owned by this struct.
-// No aliasing occurs because checkout/return ensures exclusive access per slot.
-unsafe impl Send for RegisteredBufferGroup {}
-
-// SAFETY: The atomic bitset provides thread-safe checkout/return. Buffer memory
-// is only accessed by the holder of a checked-out slot index.
-unsafe impl Sync for RegisteredBufferGroup {}
-
-pub use crate::io_uring_common::{RegisteredBufferStats, RegisteredBufferStatus};
-
-/// A checked-out buffer slot from a [`RegisteredBufferGroup`].
-///
-/// Provides access to the underlying buffer memory and the buffer index needed
-/// for `ReadFixed`/`WriteFixed` SQEs. The slot is returned to the group on drop.
-pub struct RegisteredBufferSlot<'a> {
-    group: &'a RegisteredBufferGroup,
-    /// The buffer index within the registered group (used as `buf_index` in SQEs).
-    index: u16,
-}
-
-impl<'a> RegisteredBufferSlot<'a> {
-    /// Returns the buffer index for use in `ReadFixed`/`WriteFixed` SQEs.
-    #[inline]
-    #[must_use]
-    pub fn buf_index(&self) -> u16 {
-        self.index
-    }
-
-    /// Returns a mutable pointer to the buffer memory.
-    ///
-    /// Callers must ensure writes stay within `buffer_size()` bounds.
-    #[inline]
-    #[must_use]
-    pub fn as_mut_ptr(&mut self) -> *mut u8 {
-        self.group.buffers[self.index as usize]
-    }
-
-    /// Returns a const pointer to the buffer memory.
-    #[inline]
-    #[must_use]
-    pub fn as_ptr(&self) -> *const u8 {
-        self.group.buffers[self.index as usize]
-    }
-
-    /// Returns the size of the buffer in bytes.
-    #[inline]
-    #[must_use]
-    pub fn buffer_size(&self) -> usize {
-        self.group.buffer_size
-    }
-
-    /// Returns a slice view of the buffer up to `len` bytes.
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure that `len` bytes have been initialized (e.g.,
-    /// after a successful `ReadFixed` completion with result >= `len`).
-    #[inline]
-    pub unsafe fn as_slice(&self, len: usize) -> &[u8] {
-        debug_assert!(len <= self.group.buffer_size);
-        let clamped = len.min(self.group.buffer_size);
-        unsafe { std::slice::from_raw_parts(self.as_ptr(), clamped) }
-    }
-
-    /// Returns a mutable slice view of the buffer up to `len` bytes.
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure no other references to this buffer memory exist.
-    #[inline]
-    pub unsafe fn as_mut_slice(&mut self, len: usize) -> &mut [u8] {
-        debug_assert!(len <= self.group.buffer_size);
-        let clamped = len.min(self.group.buffer_size);
-        unsafe { std::slice::from_raw_parts_mut(self.as_mut_ptr(), clamped) }
-    }
-}
-
-impl Drop for RegisteredBufferSlot<'_> {
-    fn drop(&mut self) {
-        self.group.return_slot(self.index);
-    }
-}
-
-impl RegisteredBufferGroup {
-    /// Creates a new group of page-aligned buffers and registers them with io_uring.
-    ///
-    /// Each buffer is `buffer_size` bytes, page-aligned for optimal kernel DMA.
-    /// The `count` parameter specifies how many buffers to allocate (capped at 1024).
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    /// - `count` is zero
-    /// - `buffer_size` is zero
-    /// - Memory allocation fails
-    /// - `IORING_REGISTER_BUFFERS` fails (kernel limit exceeded, seccomp, etc.)
-    pub fn new(ring: &RawIoUring, buffer_size: usize, count: usize) -> io::Result<Self> {
-        if count == 0 {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "buffer count must be > 0",
-            ));
-        }
-        if buffer_size == 0 {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "buffer size must be > 0",
-            ));
-        }
-        if count > MAX_REGISTERED_BUFFERS {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                format!("buffer count {count} exceeds kernel limit {MAX_REGISTERED_BUFFERS}"),
-            ));
-        }
-
-        let page_size = page_size();
-        let aligned_size = buffer_size.next_multiple_of(page_size);
-        let layout = Layout::from_size_align(aligned_size, page_size).map_err(|e| {
-            io::Error::new(
-                io::ErrorKind::InvalidInput,
-                format!("invalid buffer layout: {e}"),
-            )
-        })?;
-
-        let mut buffers = Vec::with_capacity(count);
-        let mut iovecs = Vec::with_capacity(count);
-
-        for _ in 0..count {
-            // Safety: layout has non-zero size and valid alignment.
-            let ptr = unsafe { alloc::alloc_zeroed(layout) };
-            if ptr.is_null() {
-                // Clean up already-allocated buffers.
-                for prev in &buffers {
-                    unsafe { alloc::dealloc(*prev, layout) };
-                }
-                return Err(io::Error::new(
-                    io::ErrorKind::OutOfMemory,
-                    "failed to allocate page-aligned buffer",
-                ));
-            }
-            iovecs.push(libc::iovec {
-                iov_base: ptr.cast::<libc::c_void>(),
-                iov_len: aligned_size,
-            });
-            buffers.push(ptr);
-        }
-
-        // Register all buffers with the kernel.
-        // Safety: iovec pointers are valid and will remain valid until
-        // unregister_buffers is called in Drop. The buffers are owned by this
-        // struct and not moved or freed until then.
-        let register_result = unsafe { ring.submitter().register_buffers(&iovecs) };
-        if let Err(e) = register_result {
-            for ptr in &buffers {
-                unsafe { alloc::dealloc(*ptr, layout) };
-            }
-            return Err(io::Error::other(format!(
-                "IORING_REGISTER_BUFFERS failed: {e}"
-            )));
-        }
-
-        // Initialize free bitset - all slots start as free (bit = 1).
-        let words = count.div_ceil(64);
-        let mut free_bitset = Vec::with_capacity(words);
-        for i in 0..words {
-            let bits_in_word = if i < words - 1 {
-                64
-            } else {
-                let remainder = count % 64;
-                if remainder == 0 { 64 } else { remainder }
-            };
-            // Set `bits_in_word` lower bits to 1.
-            let mask = if bits_in_word == 64 {
-                u64::MAX
-            } else {
-                (1u64 << bits_in_word) - 1
-            };
-            free_bitset.push(AtomicU64::new(mask));
-        }
-
-        Ok(Self {
-            buffers,
-            layout,
-            buffer_size: aligned_size,
-            count,
-            free_bitset,
-            total_acquires: AtomicU64::new(0),
-            total_misses: AtomicU64::new(0),
-        })
-    }
-
-    /// Attempts to register buffers with the given ring, returning `None` on failure.
-    ///
-    /// This is the preferred entry point for optional buffer registration - it
-    /// never returns an error, making it safe to call speculatively.
-    pub fn try_new(ring: &RawIoUring, buffer_size: usize, count: usize) -> Option<Self> {
-        Self::new(ring, buffer_size, count).ok()
-    }
-
-    /// Attempts registration honoring an `enabled` flag, returning the group
-    /// (if any) plus a [`RegisteredBufferStatus`] that distinguishes the
-    /// "disabled by config" and "registration failed" paths.
-    ///
-    /// - When `enabled` is `false`: returns `(None, Disabled)` without calling
-    ///   the kernel.
-    /// - When `enabled` is `true` and registration succeeds: returns
-    ///   `(Some(group), Enabled)`.
-    /// - When `enabled` is `true` and the kernel rejects registration:
-    ///   returns `(None, RegistrationFailed { reason })` carrying the
-    ///   formatted `errno` for telemetry.
-    pub fn try_new_with_status(
-        ring: &RawIoUring,
-        buffer_size: usize,
-        count: usize,
-        enabled: bool,
-    ) -> (Option<Self>, RegisteredBufferStatus) {
-        if !enabled {
-            return (None, RegisteredBufferStatus::Disabled);
-        }
-        match Self::new(ring, buffer_size, count) {
-            Ok(g) => (Some(g), RegisteredBufferStatus::Enabled),
-            Err(e) => (
-                None,
-                RegisteredBufferStatus::RegistrationFailed {
-                    reason: e.to_string(),
-                },
-            ),
-        }
-    }
-
-    /// Returns the number of buffers in this group.
-    #[inline]
-    #[must_use]
-    pub fn count(&self) -> usize {
-        self.count
-    }
-
-    /// Returns the size of each buffer in bytes (page-aligned).
-    #[inline]
-    #[must_use]
-    pub fn buffer_size(&self) -> usize {
-        self.buffer_size
-    }
-
-    /// Returns the number of currently available (free) buffer slots.
-    #[must_use]
-    pub fn available(&self) -> usize {
-        self.free_bitset
-            .iter()
-            .map(|word| word.load(Ordering::Relaxed).count_ones() as usize)
-            .sum()
-    }
-
-    /// Checks out a free buffer slot for use with `ReadFixed`/`WriteFixed`.
-    ///
-    /// Returns `None` if all slots are currently in use. The returned
-    /// [`RegisteredBufferSlot`] automatically returns the slot on drop.
-    ///
-    /// Bumps `total_acquires` on entry and `total_misses` on the `None`
-    /// return path. Both counters are `Relaxed` and feed [`stats`](Self::stats).
-    pub fn checkout(&self) -> Option<RegisteredBufferSlot<'_>> {
-        self.total_acquires.fetch_add(1, Ordering::Relaxed);
-        for (word_idx, word) in self.free_bitset.iter().enumerate() {
-            loop {
-                let current = word.load(Ordering::Acquire);
-                if current == 0 {
-                    break; // No free bits in this word.
-                }
-                let bit = current.trailing_zeros();
-                let mask = 1u64 << bit;
-                // CAS to claim the bit (set it to 0).
-                match word.compare_exchange_weak(
-                    current,
-                    current & !mask,
-                    Ordering::AcqRel,
-                    Ordering::Acquire,
-                ) {
-                    Ok(_) => {
-                        let index = (word_idx * 64 + bit as usize) as u16;
-                        return Some(RegisteredBufferSlot { group: self, index });
-                    }
-                    Err(_) => continue, // Retry on contention.
-                }
-            }
-        }
-        self.total_misses.fetch_add(1, Ordering::Relaxed);
-        None
-    }
-
-    /// Returns a snapshot of acquire / miss counters.
-    ///
-    /// The returned [`RegisteredBufferStats`] reads each atomic counter
-    /// independently with `Relaxed` ordering. Individual values are
-    /// accurate but the snapshot is not strictly consistent across the
-    /// two fields under concurrent updates - identical to the
-    /// `BufferPoolStats` pattern in the engine's local-copy buffer pool.
-    ///
-    /// Used by the adaptive buffer sizer to drive an EMA-smoothed miss-rate
-    /// signal. See `docs/audits/io-uring-adaptive-buffer-sizing.md` for the
-    /// design.
-    #[must_use]
-    pub fn stats(&self) -> RegisteredBufferStats {
-        RegisteredBufferStats {
-            total_acquires: self.total_acquires.load(Ordering::Relaxed),
-            total_misses: self.total_misses.load(Ordering::Relaxed),
-        }
-    }
-
-    /// Returns a buffer slot to the free pool.
-    fn return_slot(&self, index: u16) {
-        let word_idx = index as usize / 64;
-        let bit = index as usize % 64;
-        let mask = 1u64 << bit;
-        self.free_bitset[word_idx].fetch_or(mask, Ordering::Release);
-    }
-
-    /// Unregisters all buffers from the io_uring instance.
-    ///
-    /// Called by the ring owner before dropping the ring. If the ring is dropped
-    /// first, the kernel automatically unregisters buffers - but explicit
-    /// unregistration is preferred for deterministic cleanup.
-    pub fn unregister(&self, ring: &RawIoUring) -> io::Result<()> {
-        ring.submitter().unregister_buffers()
-    }
-}
-
-impl Drop for RegisteredBufferGroup {
-    fn drop(&mut self) {
-        // Drop ordering invariant (see module docs): the ring fd is closed
-        // before this Drop runs, which causes the kernel to release the
-        // pinning on these buffer pages (io_uring_register(2) /
-        // fs/io_uring.c:io_sqe_buffers_unregister). All we do here is free
-        // the user-side memory.
-        //
-        // Panic safety: alloc::dealloc does not panic when called with a
-        // matching layout, so this Drop is safe during stack unwinding.
-        // No double-panic / process abort can be triggered from here.
-        //
-        // SIGKILL: this code does not run on forced termination, but the
-        // kernel reclaims both the ring fd and the registered pages as part
-        // of process teardown.
-        for ptr in &self.buffers {
-            // Safety: each pointer was allocated with self.layout via
-            // alloc::alloc_zeroed and has not been freed yet. We own all
-            // buffers exclusively at this point - no slot can outlive the
-            // group because RegisteredBufferSlot borrows &self.
-            unsafe { alloc::dealloc(*ptr, self.layout) };
-        }
-    }
-}
-
-/// Returns the system page size.
-fn page_size() -> usize {
-    // Safety: sysconf is always safe to call with _SC_PAGESIZE.
-    let size = unsafe { libc::sysconf(libc::_SC_PAGESIZE) };
-    if size <= 0 {
-        4096 // Fallback to 4K, the most common page size.
-    } else {
-        size as usize
-    }
-}
-
-/// Submits a batch of `ReadFixed` SQEs reading into registered buffers.
-///
-/// Reads `total_len` bytes from the file starting at `base_offset`, using
-/// registered buffers from `slots`. Each slot handles one chunk of data.
-/// Completions are collected and the total bytes read is returned.
-///
-/// The `slots` parameter provides buffer indices and pointers. Callers must
-/// ensure slots are checked out from a `RegisteredBufferGroup` that is
-/// registered with the same ring.
-///
-/// Exposed publicly under `#[doc(hidden)]` so integration tests in
-/// `crates/fast_io/tests/` can exercise short-read paths (NFS, FUSE,
-/// slow block devices). Not part of the stable public API.
-#[doc(hidden)]
-pub fn submit_read_fixed_batch(
-    ring: &mut RawIoUring,
-    fd: io_uring::types::Fd,
-    output: &mut [u8],
-    base_offset: u64,
-    slots: &[RegisteredBufferSlotInfo],
-    fixed_fd_slot: i32,
-) -> io::Result<usize> {
-    use super::batching::maybe_fixed_file;
-    use io_uring::opcode::ReadFixed;
-
-    if output.is_empty() || slots.is_empty() {
-        return Ok(0);
-    }
-
-    let mut total_read = 0usize;
-    let total = output.len();
-    let chunk_size = slots[0].buffer_size;
-
-    // Process in rounds, one SQE per slot per round.
-    while total_read < total {
-        let remaining = total - total_read;
-        let n_sqes = remaining.div_ceil(chunk_size).min(slots.len());
-        let mut submitted = 0u32;
-
-        // Track how many bytes each SQE requested for short-read detection.
-        let mut requested_per_sqe: Vec<usize> = Vec::with_capacity(n_sqes);
-
-        for (i, slot) in slots.iter().enumerate().take(n_sqes) {
-            let offset_in_output = total_read + i * chunk_size;
-            let want = chunk_size.min(total - offset_in_output);
-            let file_offset = base_offset + offset_in_output as u64;
-
-            let entry = ReadFixed::new(fd, slot.ptr, want as u32, slot.buf_index)
-                .offset(file_offset)
-                .build()
-                .user_data(i as u64);
-            let entry = maybe_fixed_file(entry, fixed_fd_slot);
-
-            // Safety: the registered buffer at slot is valid and pinned for
-            // the duration of this submit_and_wait cycle.
-            unsafe {
-                ring.submission()
-                    .push(&entry)
-                    .map_err(|_| io::Error::other("submission queue full"))?;
-            }
-            requested_per_sqe.push(want);
-            submitted += 1;
-        }
-
-        if submitted == 0 {
-            break;
-        }
-
-        ring.submit_and_wait(submitted as usize)?;
-
-        // Collect actual bytes read per SQE index. CQEs may arrive out of order.
-        let mut actual_per_sqe = vec![0usize; submitted as usize];
-
-        let mut completed = 0u32;
-        while completed < submitted {
-            let cqe = ring
-                .completion()
-                .next()
-                .ok_or_else(|| io::Error::other("missing CQE"))?;
-
-            let idx = cqe.user_data() as usize;
-            let result = cqe.result();
-
-            if result < 0 {
-                return Err(io::Error::from_raw_os_error(-result));
-            }
-
-            let bytes = result as usize;
-            actual_per_sqe[idx] = bytes;
-
-            let out_start = total_read + idx * chunk_size;
-            let out_end = (out_start + bytes).min(total);
-            let copy_len = out_end - out_start;
-
-            // Safety: the kernel wrote `bytes` into the registered buffer.
-            // We copy from the registered buffer into the caller's output slice.
-            unsafe {
-                ptr::copy_nonoverlapping(
-                    slots[idx].ptr,
-                    output[out_start..].as_mut_ptr(),
-                    copy_len,
-                );
-            }
-
-            completed += 1;
-        }
-
-        // Advance by the contiguous prefix of fully-read SQEs. If SQE `i`
-        // returned fewer bytes than requested (short read - common on NFS,
-        // FUSE, and slow block devices), we stop at that point so the outer
-        // loop retries from the correct offset.
-        let mut batch_advance = 0usize;
-        for i in 0..submitted as usize {
-            batch_advance += actual_per_sqe[i];
-            if actual_per_sqe[i] < requested_per_sqe[i] {
-                break;
-            }
-        }
-
-        if batch_advance == 0 {
-            break; // EOF or zero-length read - avoid infinite loop.
-        }
-        total_read += batch_advance;
-    }
-
-    // submit_read_fixed_batch invariant: on the success path the reported
-    // byte count must never exceed the caller's output capacity. A regression
-    // here would let callers read uninitialised tail memory or write past
-    // their slice. The per-CQE copy is already clamped via `out_end.min(total)`
-    // so this guards the returned length itself.
-    let reported = total_read.min(total);
-    debug_assert!(
-        reported <= output.len(),
-        "submit_read_fixed_batch returned {reported} bytes for an output of {} bytes",
-        output.len()
-    );
-    Ok(reported)
-}
-
-/// Submits a batch of `WriteFixed` SQEs writing from registered buffers.
-///
-/// Writes `data` to the file starting at `base_offset`, copying chunks into
-/// registered buffers and submitting `WriteFixed` SQEs. Returns the total
-/// bytes written.
-pub(super) fn submit_write_fixed_batch(
-    ring: &mut RawIoUring,
-    fd: io_uring::types::Fd,
-    data: &[u8],
-    base_offset: u64,
-    slots: &[RegisteredBufferSlotInfo],
-    fixed_fd_slot: i32,
-) -> io::Result<usize> {
-    use super::batching::maybe_fixed_file;
-    use io_uring::opcode::WriteFixed;
-
-    if data.is_empty() || slots.is_empty() {
-        return Ok(0);
-    }
-
-    let total = data.len();
-    let mut total_written = 0usize;
-    let chunk_size = slots[0].buffer_size;
-
-    while total_written < total {
-        let remaining = total - total_written;
-        let n_sqes = remaining.div_ceil(chunk_size).min(slots.len());
-        let mut submitted = 0u32;
-
-        for (i, slot) in slots.iter().enumerate().take(n_sqes) {
-            let src_start = total_written + i * chunk_size;
-            let want = chunk_size.min(total - src_start);
-            let file_offset = base_offset + src_start as u64;
-
-            // Copy data into registered buffer.
-            // Safety: registered buffer at slot is valid and large enough.
-            unsafe {
-                ptr::copy_nonoverlapping(data[src_start..].as_ptr(), slot.ptr, want);
-            }
-
-            let entry = WriteFixed::new(fd, slot.ptr, want as u32, slot.buf_index)
-                .offset(file_offset)
-                .build()
-                .user_data(i as u64);
-            let entry = maybe_fixed_file(entry, fixed_fd_slot);
-
-            // Safety: the registered buffer contains valid data and is pinned
-            // for the duration of this submit_and_wait cycle.
-            unsafe {
-                ring.submission()
-                    .push(&entry)
-                    .map_err(|_| io::Error::other("submission queue full"))?;
-            }
-            submitted += 1;
-        }
-
-        if submitted == 0 {
-            break;
-        }
-
-        ring.submit_and_wait(submitted as usize)?;
-
-        let mut batch_written = 0usize;
-        let mut completed = 0u32;
-        while completed < submitted {
-            let cqe = ring
-                .completion()
-                .next()
-                .ok_or_else(|| io::Error::other("missing CQE"))?;
-
-            let result = cqe.result();
-            if result < 0 {
-                return Err(io::Error::from_raw_os_error(-result));
-            }
-            if result == 0 {
-                return Err(io::Error::new(
-                    io::ErrorKind::WriteZero,
-                    "write_fixed returned 0 bytes",
-                ));
-            }
-
-            batch_written += result as usize;
-            completed += 1;
-        }
-
-        total_written += batch_written;
-    }
-
-    Ok(total_written)
-}
-
-/// Lightweight info struct for passing registered buffer metadata to batch helpers.
-///
-/// Avoids lifetime complications of passing `RegisteredBufferSlot` references
-/// into the batch submission functions.
-///
-/// Exposed publicly under `#[doc(hidden)]` to allow integration tests to
-/// drive [`submit_read_fixed_batch`]. Not part of the stable public API.
-#[doc(hidden)]
-pub struct RegisteredBufferSlotInfo {
-    /// Raw pointer to the registered buffer memory.
-    pub ptr: *mut u8,
-    /// Buffer index for `ReadFixed`/`WriteFixed` SQEs.
-    pub buf_index: u16,
-    /// Size of the buffer in bytes.
-    pub buffer_size: usize,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn page_size_is_positive_and_power_of_two() {
-        let ps = page_size();
-        assert!(ps > 0);
-        assert!(ps.is_power_of_two());
-    }
-
-    #[test]
-    fn registered_buffer_group_rejects_zero_count() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return, // io_uring not available
-        };
-        let result = RegisteredBufferGroup::new(&ring, 4096, 0);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn registered_buffer_group_rejects_zero_size() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let result = RegisteredBufferGroup::new(&ring, 0, 4);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn registered_buffer_group_rejects_excessive_count() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let result = RegisteredBufferGroup::new(&ring, 4096, MAX_REGISTERED_BUFFERS + 1);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn registered_buffer_group_create_and_checkout() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 4) {
-            Ok(g) => g,
-            Err(_) => return, // Registration failed (seccomp, kernel limit, etc.)
-        };
-
-        assert_eq!(group.count(), 4);
-        assert!(group.buffer_size() >= 4096);
-        assert_eq!(group.available(), 4);
-
-        // Check out all 4 slots.
-        let mut s0 = group.checkout().expect("slot 0");
-        assert_eq!(group.available(), 3);
-        let s1 = group.checkout().expect("slot 1");
-        let mut s2 = group.checkout().expect("slot 2");
-        let mut s3 = group.checkout().expect("slot 3");
-        assert_eq!(group.available(), 0);
-
-        // No more slots available.
-        assert!(group.checkout().is_none());
-
-        // Return one slot.
-        drop(s1);
-        assert_eq!(group.available(), 1);
-
-        // Check out again.
-        let mut s1b = group.checkout().expect("slot 1 reacquired");
-        assert_eq!(group.available(), 0);
-
-        // Verify buffer pointers are non-null and unique.
-        let ptrs: Vec<*mut u8> = [&mut s0, &mut s1b, &mut s2, &mut s3]
-            .iter_mut()
-            .map(|s| s.as_mut_ptr())
-            .collect();
-        for p in &ptrs {
-            assert!(!p.is_null());
-        }
-        // All pointers should be distinct.
-        for i in 0..ptrs.len() {
-            for j in (i + 1)..ptrs.len() {
-                assert_ne!(ptrs[i], ptrs[j], "slots {i} and {j} share a pointer");
-            }
-        }
-
-        drop(s0);
-        drop(s1b);
-        drop(s2);
-        drop(s3);
-        assert_eq!(group.available(), 4);
-
-        // Explicit unregister.
-        let _ = group.unregister(&ring);
-    }
-
-    #[test]
-    fn buffer_slot_read_write_memory() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        let mut slot = group.checkout().expect("checkout");
-
-        // Write a pattern into the buffer.
-        let pattern = b"hello io_uring registered buffers!";
-        unsafe {
-            ptr::copy_nonoverlapping(pattern.as_ptr(), slot.as_mut_ptr(), pattern.len());
-            let read_back = slot.as_slice(pattern.len());
-            assert_eq!(read_back, pattern);
-        }
-
-        drop(slot);
-        let _ = group.unregister(&ring);
-    }
-
-    #[test]
-    fn read_fixed_write_fixed_roundtrip() {
-        let ring = match RawIoUring::new(64) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 4) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("fixed_roundtrip.bin");
-
-        // Generate test data larger than one buffer.
-        let test_data: Vec<u8> = (0..12000u32).map(|i| (i % 256) as u8).collect();
-        std::fs::write(&path, &test_data).unwrap();
-
-        // Collect slot info for batch operations.
-        let mut checked_out: Vec<_> = (0..4).filter_map(|_| group.checkout()).collect();
-        let slot_infos: Vec<RegisteredBufferSlotInfo> = checked_out
-            .iter_mut()
-            .map(|s| RegisteredBufferSlotInfo {
-                ptr: s.as_mut_ptr(),
-                buf_index: s.buf_index(),
-                buffer_size: s.buffer_size(),
-            })
-            .collect();
-
-        // Read the file using ReadFixed.
-        let file = std::fs::File::open(&path).unwrap();
-        let raw_fd = {
-            use std::os::unix::io::AsRawFd;
-            file.as_raw_fd()
-        };
-        let fd = io_uring::types::Fd(raw_fd);
-
-        let mut read_buf = vec![0u8; test_data.len()];
-        let mut ring_rw = ring;
-        let bytes_read = submit_read_fixed_batch(
-            &mut ring_rw,
-            fd,
-            &mut read_buf,
-            0,
-            &slot_infos,
-            super::super::batching::NO_FIXED_FD,
-        )
-        .unwrap();
-
-        assert_eq!(bytes_read, test_data.len());
-        assert_eq!(read_buf, test_data);
-
-        // Write using WriteFixed to a new file.
-        let write_path = dir.path().join("fixed_write_out.bin");
-        let write_file = std::fs::File::create(&write_path).unwrap();
-        let write_fd = {
-            use std::os::unix::io::AsRawFd;
-            io_uring::types::Fd(write_file.as_raw_fd())
-        };
-
-        let bytes_written = submit_write_fixed_batch(
-            &mut ring_rw,
-            write_fd,
-            &test_data,
-            0,
-            &slot_infos,
-            super::super::batching::NO_FIXED_FD,
-        )
-        .unwrap();
-
-        assert_eq!(bytes_written, test_data.len());
-        drop(write_file); // Flush.
-
-        let written_data = std::fs::read(&write_path).unwrap();
-        assert_eq!(written_data, test_data);
-
-        drop(checked_out);
-        let _ = group.unregister(&ring_rw);
-    }
-
-    /// Reads with an output buffer larger than the file to trigger a natural
-    /// short read (EOF before buffer is full). Before the fix, the function
-    /// would advance past unread bytes, returning `total` even though the
-    /// file was smaller - silently zero-filling the tail.
-    #[test]
-    fn read_fixed_batch_short_read_at_eof() {
-        let ring = match RawIoUring::new(64) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 4) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("short_read.bin");
-
-        // File is 5000 bytes but we ask to read 16384 (4 * 4096).
-        let test_data: Vec<u8> = (0..5000u32).map(|i| (i % 251) as u8).collect();
-        std::fs::write(&path, &test_data).unwrap();
-
-        let mut checked_out: Vec<_> = (0..4).filter_map(|_| group.checkout()).collect();
-        let slot_infos: Vec<RegisteredBufferSlotInfo> = checked_out
-            .iter_mut()
-            .map(|s| RegisteredBufferSlotInfo {
-                ptr: s.as_mut_ptr(),
-                buf_index: s.buf_index(),
-                buffer_size: s.buffer_size(),
-            })
-            .collect();
-
-        let file = std::fs::File::open(&path).unwrap();
-        let raw_fd = {
-            use std::os::unix::io::AsRawFd;
-            file.as_raw_fd()
-        };
-        let fd = io_uring::types::Fd(raw_fd);
-
-        // Request more bytes than the file contains.
-        let request_size = 4 * 4096;
-        let mut read_buf = vec![0xFFu8; request_size];
-        let mut ring_rw = ring;
-        let bytes_read = submit_read_fixed_batch(
-            &mut ring_rw,
-            fd,
-            &mut read_buf,
-            0,
-            &slot_infos,
-            super::super::batching::NO_FIXED_FD,
-        )
-        .unwrap();
-
-        // Must return exactly the file size, not the request size.
-        assert_eq!(bytes_read, test_data.len());
-        assert_eq!(&read_buf[..bytes_read], &test_data[..]);
-
-        drop(checked_out);
-        let _ = group.unregister(&ring_rw);
-    }
-
-    /// Reads a file that is smaller than a single registered buffer chunk.
-    /// The first SQE returns a short read (file size < chunk size), and the
-    /// function must report only the actual bytes read.
-    #[test]
-    fn read_fixed_batch_file_smaller_than_chunk() {
-        let ring = match RawIoUring::new(64) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("tiny.bin");
-
-        let test_data = b"small file content";
-        std::fs::write(&path, test_data).unwrap();
-
-        let mut checked_out: Vec<_> = (0..2).filter_map(|_| group.checkout()).collect();
-        let slot_infos: Vec<RegisteredBufferSlotInfo> = checked_out
-            .iter_mut()
-            .map(|s| RegisteredBufferSlotInfo {
-                ptr: s.as_mut_ptr(),
-                buf_index: s.buf_index(),
-                buffer_size: s.buffer_size(),
-            })
-            .collect();
-
-        let file = std::fs::File::open(&path).unwrap();
-        let raw_fd = {
-            use std::os::unix::io::AsRawFd;
-            file.as_raw_fd()
-        };
-        let fd = io_uring::types::Fd(raw_fd);
-
-        // Request 8192 bytes (2 chunks) but file is only 18 bytes.
-        let mut read_buf = vec![0xFFu8; 8192];
-        let mut ring_rw = ring;
-        let bytes_read = submit_read_fixed_batch(
-            &mut ring_rw,
-            fd,
-            &mut read_buf,
-            0,
-            &slot_infos,
-            super::super::batching::NO_FIXED_FD,
-        )
-        .unwrap();
-
-        assert_eq!(bytes_read, test_data.len());
-        assert_eq!(&read_buf[..bytes_read], &test_data[..]);
-
-        drop(checked_out);
-        let _ = group.unregister(&ring_rw);
-    }
-
-    /// Drop ordering invariant: dropping the `RegisteredBufferGroup` BEFORE
-    /// the `RawIoUring` is sound. The kernel still holds the buffer pinning
-    /// (released later when the ring fd closes), but we may safely deallocate
-    /// the user-side memory because `Drop` does not touch the ring.
-    #[test]
-    fn drop_group_before_ring_does_not_panic() {
-        let mut ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        // Drop the group first while the ring is still alive.
-        drop(group);
-
-        // Ring is still usable for ordinary operations after the group dies.
-        // Submitting a no-op (nop) verifies the ring fd remains valid.
-        let entry = io_uring::opcode::Nop::new().build().user_data(0xdead);
-        // Safety: SQE is a Nop with no buffer pointers.
-        unsafe {
-            ring.submission()
-                .push(&entry)
-                .expect("nop submission after group drop");
-        }
-        ring.submit_and_wait(1).expect("nop completes");
-        let cqe = ring.completion().next().expect("nop CQE");
-        assert_eq!(cqe.user_data(), 0xdead);
-        assert_eq!(cqe.result(), 0);
-    }
-
-    /// Drop ordering invariant: dropping the ring BEFORE the group is the
-    /// natural order used by `IoUringReader`/`IoUringWriter`. The kernel
-    /// auto-releases the buffer pinning when the ring fd closes; the group
-    /// then frees user-side memory in its own Drop.
-    #[test]
-    fn drop_ring_before_group_frees_memory_cleanly() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        // Close the ring first - kernel releases buffer pinning.
-        drop(ring);
-
-        // Now dropping the group must still be sound: it deallocates user
-        // memory and never accesses the (now-closed) ring fd.
-        drop(group);
-    }
-
-    /// Mirrors the field declaration order used by `IoUringReader` and
-    /// `IoUringWriter`: ring before registered_buffers. Verifies that
-    /// implicit drop runs ring-first then group-second without aborting.
-    #[test]
-    fn struct_field_drop_order_matches_callers() {
-        struct OwnerLikeReader {
-            #[allow(dead_code)]
-            ring: RawIoUring,
-            #[allow(dead_code)]
-            registered_buffers: Option<RegisteredBufferGroup>,
-        }
-
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = RegisteredBufferGroup::try_new(&ring, 4096, 2);
-
-        let owner = OwnerLikeReader {
-            ring,
-            registered_buffers: group,
-        };
-
-        // Implicit drop in declaration order: ring first, then group.
-        // Must complete without panic or process abort.
-        drop(owner);
-    }
-
-    /// Panic during slot use must not corrupt the group: dropping the slot
-    /// during unwinding returns it to the free list, and dropping the group
-    /// during unwinding deallocates buffers safely.
-    #[test]
-    fn panic_during_slot_use_unwinds_cleanly() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let _slot = group.checkout().expect("slot checkout");
-            // Slot is held; panic forces Drop during unwinding.
-            panic!("simulated panic during slot use");
-        }));
-
-        assert!(result.is_err(), "panic should propagate via catch_unwind");
-
-        // Slot must have been returned to the free list during unwinding.
-        assert_eq!(
-            group.available(),
-            2,
-            "slot should be returned on panic-driven drop"
-        );
-
-        // Group is still usable after the panic.
-        let _slot_again = group.checkout().expect("re-checkout after panic");
-    }
-
-    /// `unregister()` returns an error when the buffer set has already been
-    /// released by closing the ring (or never registered). The error must
-    /// be reported to the caller; it must NOT cause a panic or abort.
-    #[test]
-    fn unregister_after_ring_closed_returns_error_or_ok() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        // Successful explicit unregister against the live ring.
-        let first = group.unregister(&ring);
-        assert!(
-            first.is_ok(),
-            "first unregister against live ring should succeed: {first:?}"
-        );
-
-        // A second unregister has nothing to release; the kernel may return
-        // EINVAL/ENXIO. The wrapper must surface this gracefully (Result),
-        // never panic. The exact error code is kernel-dependent, so we just
-        // require the call returns (Ok or Err) without panicking.
-        let _ = group.unregister(&ring);
-    }
-
-    /// User-side buffer memory must be freed regardless of whether
-    /// `unregister()` was called. We verify this by exercising both code
-    /// paths (with and without explicit unregister) and confirming Drop
-    /// completes without panic - leak detection is delegated to ASan/Miri
-    /// in CI when available.
-    #[test]
-    fn buffers_freed_with_or_without_explicit_unregister() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-
-        // Path A: explicit unregister, then drop.
-        {
-            let group = match RegisteredBufferGroup::new(&ring, 4096, 4) {
-                Ok(g) => g,
-                Err(_) => return,
-            };
-            let _ = group.unregister(&ring);
-            drop(group);
-        }
-
-        // Path B: drop without explicit unregister (relies on kernel cleanup
-        // when ring closes; here we keep the ring alive across drop).
-        {
-            let group = match RegisteredBufferGroup::new(&ring, 4096, 4) {
-                Ok(g) => g,
-                Err(_) => return,
-            };
-            drop(group);
-        }
-
-        // Path C: re-register on the same ring, drop, repeat. Verifies the
-        // ring remains in a clean state for further registrations.
-        for _ in 0..3 {
-            if let Some(group) = RegisteredBufferGroup::try_new(&ring, 4096, 2) {
-                let _ = group.unregister(&ring);
-            }
-        }
-    }
-
-    /// A freshly created group reports zero acquires and zero misses.
-    #[test]
-    fn stats_initially_zero() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-        let stats = group.stats();
-        assert_eq!(stats.total_acquires, 0);
-        assert_eq!(stats.total_misses, 0);
-        assert_eq!(stats.miss_rate(), 0.0);
-    }
-
-    /// Successful checkouts bump `total_acquires` but not `total_misses`.
-    #[test]
-    fn stats_count_successful_checkouts() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 4) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        let s0 = group.checkout().expect("slot 0");
-        let s1 = group.checkout().expect("slot 1");
-        let stats = group.stats();
-        assert_eq!(stats.total_acquires, 2);
-        assert_eq!(stats.total_misses, 0);
-        assert_eq!(stats.miss_rate(), 0.0);
-
-        drop(s0);
-        drop(s1);
-    }
-
-    /// `checkout` returning `None` increments both `total_acquires` and
-    /// `total_misses`, and `miss_rate` reflects the ratio.
-    #[test]
-    fn stats_count_misses_on_exhaustion() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        // Exhaust the pool.
-        let _s0 = group.checkout().expect("slot 0");
-        let _s1 = group.checkout().expect("slot 1");
-
-        // Three forced misses.
-        assert!(group.checkout().is_none());
-        assert!(group.checkout().is_none());
-        assert!(group.checkout().is_none());
-
-        let stats = group.stats();
-        assert_eq!(stats.total_acquires, 5);
-        assert_eq!(stats.total_misses, 3);
-        let mr = stats.miss_rate();
-        assert!(
-            (mr - 3.0 / 5.0).abs() < 1e-12,
-            "expected miss_rate=0.6, got {mr}"
-        );
-    }
-
-    /// Returning a slot does not affect telemetry counters: `total_acquires`
-    /// is the lifetime acquire count, never decremented.
-    #[test]
-    fn stats_not_decremented_on_return() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        let s = group.checkout().expect("slot");
-        drop(s);
-        assert_eq!(group.stats().total_acquires, 1);
-        assert_eq!(group.stats().total_misses, 0);
-
-        // Re-acquire the same slot bumps acquires again.
-        let s = group.checkout().expect("slot reacquired");
-        assert_eq!(group.stats().total_acquires, 2);
-        drop(s);
-    }
-
-    /// `RegisteredBufferStats::miss_rate` returns 0.0 when no acquires have
-    /// been recorded, matching the `BufferPoolStats::hit_rate` convention.
-    #[test]
-    fn stats_miss_rate_zero_when_no_acquires() {
-        let s = RegisteredBufferStats {
-            total_acquires: 0,
-            total_misses: 0,
-        };
-        assert_eq!(s.miss_rate(), 0.0);
-    }
-
-    /// `miss_rate` is exactly 1.0 when every acquire missed.
-    #[test]
-    fn stats_miss_rate_all_misses() {
-        let s = RegisteredBufferStats {
-            total_acquires: 7,
-            total_misses: 7,
-        };
-        assert!((s.miss_rate() - 1.0).abs() < 1e-12);
-    }
-
-    // Constrained-environment Drop coverage.
-    //
-    // The fixed-buffer invariants audit (PR #4022, task #2118) documents that
-    // `RegisteredBufferGroup::Drop` deliberately does NOT issue
-    // `IORING_UNREGISTER_BUFFERS`; the kernel reclaims the pinning when the
-    // ring fd closes. The tests below exercise that contract under conditions
-    // where a naive implementation would either leak userspace memory or
-    // surprise callers by silently mutating kernel state.
-
-    /// Structural proof that `Drop` does NOT call `IORING_UNREGISTER_BUFFERS`:
-    /// after dropping the group while the ring fd remains open, the kernel
-    /// must still consider buffers registered. A second `register_buffers`
-    /// call on the same ring is rejected (typically with `EBUSY`) precisely
-    /// because the prior registration is still live. After an explicit
-    /// unregister against the same ring fd, a fresh registration succeeds,
-    /// proving the ring itself is not poisoned by the silent-Drop policy.
-    #[test]
-    fn drop_does_not_release_kernel_registration() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        // Drop the group while keeping the ring alive. If Drop secretly
-        // called unregister_buffers, the kernel slot table would now be
-        // empty and the next registration would succeed.
-        drop(group);
-
-        // Attempt a fresh registration on the same ring. The kernel still
-        // holds the prior pinning, so this must be rejected. We do not
-        // assert a specific errno (kernels differ); the contract is only
-        // that the call returns an `Err`, never panics, and leaves the
-        // ring usable.
-        let second = RegisteredBufferGroup::new(&ring, 4096, 2);
-        if second.is_ok() {
-            // Some kernels (5.13+) support replace-style update semantics
-            // on a second register; in that case our invariant cannot be
-            // probed by this method. Skip rather than false-fail.
-            return;
-        }
-
-        // Release the kernel-side pinning explicitly via the submitter on
-        // the live ring fd. After this, fresh registration must succeed -
-        // confirming the ring itself is not poisoned by the silent-Drop
-        // policy, only the prior kernel-side registration was still live.
-        let _ = ring.submitter().unregister_buffers();
-        let third = RegisteredBufferGroup::new(&ring, 4096, 2);
-        assert!(
-            third.is_ok(),
-            "after explicit unregister, the ring must accept a fresh registration"
-        );
-    }
-
-    /// Drop a group whose internal tracking state is non-default: some
-    /// slots have been checked out (and returned), some checkouts have
-    /// missed, and the stats counters are non-zero. Memory must be freed
-    /// cleanly regardless of the tracking-state snapshot at drop time.
-    ///
-    /// This guards against a regression where Drop assumes a pristine
-    /// bitset / counter state. The audit (PR #4022, task #2118) notes the
-    /// invariant that `Drop` is "panic-safe" because it only calls
-    /// `alloc::dealloc`; we make the in-use-tracking case explicit here.
-    #[test]
-    fn drop_with_in_use_tracking_state_is_clean() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 4) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        // Drive the group through a realistic mid-use trajectory: acquire
-        // and release multiple times, then exhaust the pool to bump misses,
-        // then return everything so the bitset is "all free" but counters
-        // are non-zero.
-        for _ in 0..3 {
-            let s = group.checkout().expect("slot");
-            drop(s);
-        }
-        let hold: Vec<_> = (0..4).filter_map(|_| group.checkout()).collect();
-        for _ in 0..5 {
-            assert!(group.checkout().is_none());
-        }
-        let stats_before = group.stats();
-        assert!(stats_before.total_acquires >= 12);
-        assert!(stats_before.total_misses >= 5);
-        drop(hold);
-        assert_eq!(group.available(), 4);
-
-        // Slots must be dropped before the group (compile-time invariant via
-        // `RegisteredBufferSlot<'a>` borrow on `&self`). With slots released,
-        // drop the group: this must complete without panic or abort even
-        // though acquire / miss counters carry stale state from earlier
-        // activity.
-        drop(group);
-    }
-
-    /// `RegisteredBufferStats` is `Copy`, so a snapshot taken before the
-    /// group is dropped remains usable after. This documents that
-    /// telemetry consumers (e.g., the adaptive sizer) do not need to
-    /// outlive their group via lifetime coupling.
-    #[test]
-    fn stats_snapshot_survives_group_drop() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let group = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        // Generate measurable telemetry: one hit, one forced miss.
-        let s = group.checkout().expect("slot");
-        let s2 = group.checkout().expect("second slot");
-        assert!(group.checkout().is_none());
-        drop(s);
-        drop(s2);
-
-        let snapshot = group.stats();
-        assert_eq!(snapshot.total_acquires, 3);
-        assert_eq!(snapshot.total_misses, 1);
-
-        // Drop the group. The snapshot is plain integers and must remain
-        // observably identical after the source group is gone.
-        drop(group);
-
-        assert_eq!(snapshot.total_acquires, 3);
-        assert_eq!(snapshot.total_misses, 1);
-        let mr = snapshot.miss_rate();
-        assert!(
-            (mr - 1.0 / 3.0).abs() < 1e-12,
-            "miss_rate snapshot drift: {mr}"
-        );
-    }
-
-    /// When `RegisteredBufferGroup::new` fails after the first registration
-    /// is already live on the ring, the recovery path must (a) deallocate
-    /// the partially-built buffer set, (b) return an `Err` to the caller,
-    /// and (c) leave the ring in a state where a subsequent `try_new`
-    /// against a freshly unregistered ring succeeds. This is the
-    /// "early return from registration failure recovery" constrained
-    /// environment called out in task #1678.
-    #[test]
-    fn drop_on_construction_failure_does_not_double_register() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-
-        // First registration claims the ring's slot table.
-        let first = match RegisteredBufferGroup::new(&ring, 4096, 2) {
-            Ok(g) => g,
-            Err(_) => return,
-        };
-
-        // Second `new` against the same live ring is the recovery path we
-        // want to exercise. On kernels that refuse double-registration the
-        // partially-allocated buffers must be freed inside `new` before
-        // the error propagates - any leak shows up as an ASan / Miri
-        // failure. On newer kernels (5.13+) that accept replace-style
-        // update semantics the second `new` succeeds; in that case we
-        // cannot probe the failure-recovery branch and skip out cleanly.
-        let second = RegisteredBufferGroup::new(&ring, 4096, 2);
-        if second.is_ok() {
-            return;
-        }
-        drop(second);
-
-        // The first group is still functional - registration-failure
-        // recovery in the second call did not poison its state.
-        assert_eq!(first.available(), 2);
-        let s = first.checkout().expect("first group still usable");
-        drop(s);
-
-        // Explicit unregister on the live group clears kernel state.
-        let _ = first.unregister(&ring);
-        drop(first);
-
-        // After cleanup, a fresh group constructs cleanly. This proves the
-        // failure-recovery branch did not leave dangling kernel state.
-        let third = RegisteredBufferGroup::new(&ring, 4096, 2);
-        assert!(
-            third.is_ok(),
-            "ring must be reusable after a failed registration was cleaned up"
-        );
-    }
-
-    /// `try_new_with_status` with `enabled=false` returns `Disabled` without
-    /// calling the kernel - distinct from a `RegistrationFailed` outcome.
-    #[test]
-    fn try_new_with_status_disabled_when_flag_off() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let (group, status) = RegisteredBufferGroup::try_new_with_status(&ring, 4096, 4, false);
-        assert!(group.is_none());
-        assert_eq!(status, RegisteredBufferStatus::Disabled);
-        assert!(status.is_disabled() && !status.is_enabled() && !status.is_registration_failed());
-    }
-
-    /// Successful registration yields `Enabled` and a live group; constrained
-    /// environments that reject registration still produce a non-`Disabled`
-    /// status, exercising the failure branch.
-    #[test]
-    fn try_new_with_status_enabled_on_success() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let (group, status) = RegisteredBufferGroup::try_new_with_status(&ring, 4096, 4, true);
-        assert_ne!(status, RegisteredBufferStatus::Disabled);
-        if let RegisteredBufferStatus::Enabled = status {
-            assert_eq!(group.expect("Enabled implies a group").count(), 4);
-        }
-    }
-
-    /// When registration fails the status carries the formatted `errno` for
-    /// telemetry and the group is `None`. Forcing failure via the wrapper's
-    /// own `MAX_REGISTERED_BUFFERS` ceiling keeps the test portable.
-    #[test]
-    fn try_new_with_status_registration_failed_carries_reason() {
-        let ring = match RawIoUring::new(4) {
-            Ok(r) => r,
-            Err(_) => return,
-        };
-        let (group, status) = RegisteredBufferGroup::try_new_with_status(
-            &ring,
-            4096,
-            MAX_REGISTERED_BUFFERS + 1,
-            true,
-        );
-        assert!(group.is_none());
-        match status {
-            RegisteredBufferStatus::RegistrationFailed { reason } => {
-                assert!(!reason.is_empty(), "failure reason must be populated");
-            }
-            other => panic!("expected RegistrationFailed, got {other:?}"),
-        }
-    }
-}
